Implement ZIOAppPlatformSpecific (#5697)
* implement zioappplatformspecific

* fix warnings
